webapp/requirements.txt
click==7.1.2
Flask==1.1.2
greenlet==1.0.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
protobuf==3.15.5
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.4.2
Werkzeug==1.0.1
